pseudocoded.
